Add components for Criterias
